+++ src/ZenonCli/Commands/Stake.cs
@@ -10,10 +10,10 @@ namespace ZenonCli.Commands
         public class List : WalletAndConnectionCommand
         {
             [Value(0, Default = 0, MetaName = "pageIndex")]
-            public uint? PageIndex { get; set; }
+            public int? PageIndex { get; set; }
 
             [Value(1, Default = 25, MetaName = "PageSize")]
-            public uint? PageSize { get; set; }
+            public int? PageSize { get; set; }
 
             protected override async Task ProcessAsync()
             {
@@ -29,7 +29,7 @@ namespace ZenonCli.Commands
                 AssertPageRange(PageIndex.Value, PageSize.Value);
 
                 var stakeList = await Zdk!.Embedded.Stake.GetEntriesByAddress(
-                    address, PageIndex.Value, PageSize.Value);
+                    address, (uint)PageIndex.Value, (uint)PageSize.Value);
 
                 if (stakeList.Count > 0)
                 {
